Use unwrap_phase from scikit-image library for 2D angle unwrap
The phase gradient maps are proportional to the angle of the inverse
fourier transform. After taking the angle of the 2D inverse FT, it needs
to be unwrapped, which is not straightforward in 2D. This uses
unwrap_phase from the scikit-image library.

Also renames _inverse_fourier_transform to
_reconstruct_phase_gradient_maps_from_cropped_centered_diffraction_FTs
to better reflect what it does.