phasics_density_analysis: include edges of gradient maps in Baffou method
Previously, we did not equate values on the edges of the phase gradient
maps to finite differences in the to-be-solved phase map, because the
finite difference fell off the map.

For one thing, that led to the corners of the to-be-solved phase map not
to appear in any equations.

This commit adds equations equating all pixels of all gradient maps to
differences in the to-be-solved phase map, including the edges.