phasics_density_analysis: specify 0 value for corner phase instead of mean phase
Previously the choice was made to find a solution with mean phase 0. Note
that this is arbitrary - the solution is invariant under a phase offset.

This commit changes that choice to be that the upper left corner pixel
has phase 0. The reason to do that is so that A.T @ A stays sparse in the
sp_solve version of solving the phase map. And it's nice that the "background"
phase is 0.