Give final and intermediate results correct units.
Make sure the magnitude and units of the wavefront, and wavefront
gradients, are correct.

For the Baffou method, also update the magnitude of the finite
difference to reflect the actual gradient.

Also changes the coefficient for the corner phase to be in the same
order of magnitude as the rest of the array.