Correct camera resolution and consequentially the diffraction spot centers
The resolution of Phasics phase retrieval images were 4 * 4.74 um because
Phasics subsamples 4x4. The camera resolution of the raw camera images
is 4.74 um.

This means that, since the locations of diffraction spot centers are now
given by spatial frequency quantities, they need to be updated as well.

The GRATING_CAMERA_DISTANCE, which is determined by fitting our phase
results to Phasics's phase results, is also updated, since the previous 
fit was done with the wrong value of CAMERA_RESOLUTION.