Interpolate cropped FT image for recentering.
After cropping the FT image on a diffraction spot center, the image is
translated so that the diffraction spot center becomes the new origin.
Previously that involved a np.roll() because the translation was done by
integral pixels.

Since diffraction spot centers no longer need to be at integral pixel
locations, use interpolation to center the diffraction spot center at
the origin.