Make locate_diffraction_spots() a utility and drop from analysis.
Previously, _locate_diffraction_spots() was a placeholder that just
returned specific values for diffraction spot centers, with the intention
to someday write a function that would calculate them from the FT image.

Instead, the diffraction spot centers have become a class attribute.

_locate_diffraction_spots is now no longer part of the analysis, but
it is replaced by locate_diffraction_spots, which is a utility function
to calculate the fractional locations of diffraction spot centers by
2D parabolic interpolation.