Updated Calibration to be a Set Factor
Calibrating off of the picoscope charge only works when the beam isn't clipped or saturated.  So I now find a constant calibration factor and use that for all camera images with the same exposure and trigger delay.  Currently just hardcoded constants in my modules file, but this should be updated to something more elegant in the future.