first commit of Spot Alignment for HDR
This program takes 2 or more folders of images at varying saturation levels and gives the user the option to either align all the images to the center of their frame, or find the images that match the best with each other and only saves those (also aligned). Processing includes bkg sub, alignment, cropping of image frame.