developed functional implementation of U_HighResMagSpec.py
using the the "python_delgate.py" function, a functional implementation of online image analysis is complete. On the labview side, there is no user input needed. The python analysis takes the device name as an input. That is then used in "image_wrapper_functions.py" (should be renamed) to execute a specific analysis.