Update image_wrapper_functions.py
added "LabviewGEECSImageAnalyzer" which becomes a base class for online image analysis using the GEECs driver. it should only handle things specific to the labview implementation. One method added so far is a pre-processing step that makes use of the background image produced by labview. This can be used or ignored. I would like to add another method in this base class to "simulate" images to help expedite development of new analysis routines. Using a simulate case allows for full testing of the labview/python integration and gives the use access to using the existing camera GUI for inspecting the processed images to make sure things are as they should be.