Update dependencies such that Py3.9 and Py3.7 can both install
Updates pyproject.toml and poetry.lock files for GEECS-PythonAPI,
ImageAnalysis, and PhasicsDensity.

The dependencies are slightly different than the former
integrate_image_analysis_python3.7 branch, because in that one the
version for scikit-image was incorrectly specified as "^0.19" which should
be ">=0.19,<1.0"

GEECS-PythonAPI requires Python3.9, and thus its lockfile has
scikit-image 0.21, numpy 0.25, and scipy 1.11, whereas ImageAnalysis
and PhasicsDensity, which must support Python3.7, have numpy 1.21,
scipy 1.7, and scikit-image 0.19.