BREAKING change: phasics_density_analysis: add negative sign in retrieving angle
The wavefronts that were produced by the algorithm were always of the
wrong sign. This commit corrects that.

The sign is inserted in the part that retrieves the angle of the inverse
Fourier transform. According to Primot (ref [2] in phasics_density_analysis.py)
the interferogram should be proportional to cos(2*pi/d * (x - z dg/dx)),
so after centering, the inverse Fourier transform should be 
A(x) * exp(-i 2*pi/d z dg/dx)

As a result, scripts that use this algorithm need to add or remove a
minus sign.