phasics_density_analysis: do frequency-centering in real space
After cropping a diffraction spot in Fourier space, the spot is then
translated to the center of the Fourier domain, and then the inverse FT
is taken. This necessarily involves interpolation, which distorts the
spot because the real and imaginary components change rapidly. This seems
to result in a large change in phase near y=0 and near x=0 in real space,
in turn dividing the wavefront gradient into quadrants.

This commit changes this sequence into taking the inverse FT of the
cropped spot without translation, then frequency-shifting in real space,
which means multiplying by exp(-i*2*pi*(nu . x)). This gives much cleaner
results, likely by avoiding interpolation, and is probably faster too.