Put central UnitRegistry in image_analyzer top-level, and pass to PhasicsImageAnalyzer
To allow both LivePostProcessing to set Quantity parameters for
PhasicsImageAnalyzer (grating period, camera resolution etc.), the
UnitRegistry has to be shared.

This puts the UnitRegistry centrally in image_analyzer.__init__, and
LivePostProcessing can pull it from there. This is better than
LivePostProcessing pulling it from PhasicsDensityAnalyzer, because other
image analyzers might need the UnitRegistry as well.

So now UC_PhasicsFileCopyImageAnalyzer needs to pass its UnitRegistry
down to PhasicsImageAnalyzer, so the latter now has a keyword option to
use an external UnitRegistry (or default to its own)