Refactored location of energy axis calibrations.
The mag spec energy axis file was separated into two parts.  The function that returns the energy axis from an array of pixels is still in the online analysis folder and gets seen by labview and the analysis modules.  The calibration script that was part of __main__ was moved to the calibration_scripts folder, which is renamed from post_analysis.