Make time_step an integer divisor of time_delay_step if possible.
Many of the algorithm steps can avoid interpolation and therefore be
faster and more accurate if time_step is an integer divisor of
time_delay_step*, such that `t - τ` is also in the `t` array.

(time_delay_step is fixed by the input image, and time_step is free to
choose but should be short enough such that most of the bandwidth of
the pulse can be resolved (Nyquist frequency >= max angular frequency))

The time step is set to the largest integer divisor of time_delay_step
if it is smaller than a few times the time_delay_step. It's a tradeoff
between
* setting the time_step to maximum_time_step and reducing the length
  of self.t and thus the size of E_sig_tτ and E_sig_ωτ,
* coercing time_step to time_delay_step to make use of faster algorithms
What that factor should be is TBD but is currently set to 5.