Pad E only to necessary length in Z and dZ/dE computation
Previously, in calculation of Z and dZ/dE, E was padded to len(self.t),
aka E_sig_tτ.shape[0], so that E_padded * E_shifted and self.E_sig_tτ[:, j]
would have the same length. But a large portion of these arrays are 0s.

This commit pads E only enough to accommodate the maximum shift by τ,
and slices self.E_sig_tτ[:, j] to this length as well.

This reduces the computation time for Z and dZ/dE