Refactored UndulatorExitCam's Analyzer
UndulatorExitCam is now structured in the same way as the MagSpecCam's analyzer.  The main analysis procedure is in the class directly, and the class can be initialized through a config file.  A default config file was provided for UndulatorExitCam.  labview_adapters.py has been updated to reflect these changes.  Lastly, the default config files were moved to a folder outside the ImageAnalysis so that it is external to the package, but within GEECS-Plugins so that it is still in the repository.