Perform stats on analyzer return keys rather than index of .values()
It is dangerous to assume the order of dictionary items, as was
happening when calling analysis['analyzer_return_dictionary'].values()
and then getting items by an index. The items or their order returned by
a function from a different library can change unexpectedly (and wasn't
even guaranteed until Python 3.7).

This commit:
* Keeps the analysis['analyzer_return_dictionary'] in dict form
* renames stats to analyzer_returns (to be more descriptive)
* Loops over analyzer_return_dictionary, name-in-hres_stats pairs