changes to geecs_device
Added a flag to the tcp subscriber state which gets switched to True on each call of handle subscription. This can be toggled externally after a state call to know if the value has been updated between calls. Could be done more elegantly. Or, figure out how to make shot number a part of the state. It's in the header of all TCP messages says Tony.

Also, added a flag about whether or not to use the alias. The tcp subscriber can take the variable name, but switches the dict to use the alias if it exists which can make it confusing if the variable name was specified