geecs device adding updating of shot number to the device state
I can see that the shot number is being read from teh tcp message in the handle_subscription method. But, it doesn't seem to be updating in the state. I assume it should be added to the generic_vars. wasn't clear to me how that is supposed to work, so I initialized it with self.state and that gets updated in handle_subscription