Add skeleton for geecs environmment and interface, plus symlink instructions
The objective here is to create geecs Badger-Plugins (environment and
interface) and prevent:
* copying the Badger-Plugins into this repo
* writing the geecs plugins in the Badger-Plugins (which means they wouldn't
  live in this repo)

My best solution is to add instructions to symlink from Badger-Plugins
to this repo's geecs environment and interface folders.

The geecs environment and interface are copies of "myenv" and "default"
respectively, so they need to be filled in.