Created a parent geecs environment class
This parent directly inherits all of the required functions of the badger environment subclass, but implements them in a general-geecs way.  This lets other environments be subclass of this parent class, and need to only specify the bare minimum.

By default, the parent class just takes the mean of whatever single observable it is given.