Added some flexibility in loading the analyzer config
There was an issue during run where the python analysis could not find the config using the relative path when it was run through LabView, and so I needed the Z: drive version.  But, when developing code locally the config file does not exist on the Z: drive yet.  Now, my current version checks for both (prioritizing the Z drive version).

NOTE:  now it becomes slightly annoying if you are developing locally and have to edit the Z drive config to test changes.  May add an override sometime later on...