Added super class function to build return dict
Now that the formatting of the return dictionary is very specific, a function like this to avoid labview and labview_adapters.py errors is necessary.

Need to test if what exactly return_lineouts needs to be for labview to not complain.  A 2x2 array of zeros is fine, while a 1x1 array of a zero is not.  How about 1x2?