Pre-compute background for U_PhasicsFileCopy imageanalyzer
The U_PhasicsFileCopy analyzer computes backgrounds from interferograms
a given file or folder and caches them. But because the image analysis
jobs run in a process pool, the cached background isn't available to
any of them, or even to the next analyze_scan() call.

This is a temporary solution to cache a background before analyzing
all the images, but it's not a great general solution, and it's not
parallelized.

Also adds _initialize_qwlsi_image_analyzer to calculate_background_from_path,
because previously initialization was only done in calculate_wavefront.

TODO: find better solution for handling Phasics background, and/or
backgrounds more generally