Pull request revisions complete but not ready for merge.
Pull request revisions made:
1) I rewrote the construct_spectral_correction() method into a get_spectral_correction_data() method which is called in the __init__(). This function reads in each specified data file, determines a common wavelength range and all unique wavelengths within this range. It then interpolates each data file over the common wavelenghts to generate a total response (or efficiency) curve. Best of all, it is performed in the init, so only once rather than for each shot.
2) Improved import via relative path.
3, 4, 5) Added documentations, descriptive naming, and more appropriate default values.

At this point, the code is likely ready to push but I would like to check functionality and benchmark execution time via framework Chris sent me prior to merging.

Additionally, Chris and I discussed the issue of floating 0th order modes on the spectrometer and how this is not currently corrected for in the analysis. As this is effectively another form of spectrum correction, I will add this functionality prior to merging this work. Plan is to use existing beam analysis class/functions to perform this.