Fixed an error with calculating charge density
For the HiResMagCam, I was just multiplying by the transverse pixel calibration factor of 43, which was giving fairly correct values for the charge density.  This was just because it is close to the inverse of the deltaE, so it was just lucky that it seemed ok.  This was not true of the ACaveMagCam, so this flawed calculation was quickly spotted.

I am now taking the sum of each vertical slice and dividing it by the array of MeV/pixel, calculated through np.ediff1d