QuadAnalysis.data_dict: only keep setpoints and indexes for which there's analysis
It's possible that for a setpoint step, analysis is not saved (user intervention,
no suitable images etc.)
Then in the QuadAnalysis data_dict, 'analyses' could contain fewer analyses
than setpoint steps, but 'setpoints' and 'indexes' still listed all the
setpoints and indexes.
This led to mismatches during rendering figures.

This commit lists only the setpoints and indexes arrays to match the
analyses.