Added an action manager to handle sequences of actions
A set of named actions that execute a sequence of commands is now availble. for example, 'insert_pmq' can be defined to shutter laser, verify laser is shuttered, move in pmq, etc.

also added reset and reinitialize methods to device_manager class so that it can be recylcled for multiple sequential scans of different types.