Substantial refactoring of datalogger and new features added to datainterface
Broke up datalogger into two separate classes: DataLogger and ScanManager. ScanManager handles everything related to setting devices, doing prescan action, etc. The basic action now is scan_manager.start_scan_thread(scan_config). This method also launches the console logging and does the most of the file writing. The file writing should probably migrate to datainterface. For testing this class, the save_data flag can be turned off and no data logging will happen

datalogger (which will be renamed soon) only handles setting up and acquiring of scalar data from the requisite devices. It can be started and stopped very simply by just invoking the data_logger method start_logging() and stop_logging(). This handles both the setting up and closing of all necessary threads.

Some new methods in DataInterface can now be used to rename files saved by Point Grey Cameras using the embedded timestamp in the image and comparing against the row in the sfile. This was written in a way that should extend to other device types that may have timestamps embedded in the saved files in different ways. Right now, the renaming is not super fast, it takes about 0.1 seconds to find the proper row and 0.1 seconds to rename the file. So, for 100 files, that's ~20 seconds. Not sure how to speed this up much more, so it probably makes sense to send this process off to execute in its own thread at the end of the scan. Right now I don't think it's threaded