added an improved sounds player class that should work across platforms
NOTE: for Mac, an additional dependency (simpleaudio) is needed. Wasn't added to the dependency list though.

The sounds are triggered in the _log_device_data method when new entries are created. In order to avoid latency issues slowing down the core code, the sound player executes in its own thread with a queue. The thread is killed in the stop_logging method