added some code to data_acquisition module to exploit existing, but previously unused, method for creating warnings if synchronous devices are not updating
currently, the method just logs a warning. For each device, a "last log time" is logged. Additional code in the scan_manager calculates "idle_time" to account for time required to set the scan parameters. The idle time is added to the last log time when calculating if the time between acquisitions is above the target threshold.