Implemented adjusting config file for experiment
There is still a lot of errors when it comes to switching to other experiments, but the base functionality is there for HTU's purposes.  IE:
if you have it set up then there is not problem.
If you don't have it at all then you can now create it through the application
You can click a button to write directly to the config if you so choose