Undulator blue diode calibration
Completed blue diode alignment calibration script for undulator screens.

The script is written so it can be executed as a standalone script but also such that it could be implemented as an analysis in a scan set up for scripted calibration scans. In the latter case, it would also be good to implement an automated  edit of the relevant config values.

I included 'Blue Centroid X' and 'Blue Centroid Y' parameter values in camera_analysis_settings.yaml for each UC_VisaEBeam device.