Created MultiScanAnalysis and MultiScanAnalysis_Visa
Created base MultiScanAnalysis class which has framework to parse and organize scan data for a list of input scans.

Created child MultiScanAnalysis_Visa class which builds upon the parent class. The child class can either perform the binning analysis of VisaEBeamAnalysis or load in processed images from a previous analysis. Then, a matrix of averaged and processed images is constructed and saved to the analysis folder for the first specified scan.

Note the processed image loading option currently does not work as the image saving function was not up to date with master. This should be troubleshot after merging.