Update scan_analysis.py
Largely cleaned up CameraImageAnalysis and VisaEBeamAnalysis.

Reorganized some class attributes in init. Consolidated path variables into a dictionary named path_dict for better organization. Extract and define the camera analysis settings as a class attribute in init.

Restructured much of the run_analysis and image processing for better streamlining and less redundant code.

Reorganized binned data into a dictionary and edited corresponding functions to properly handle the reorganized data. The main motivation for this was to better keep track of corresponding binned parameter values and data. NOTE: After talking to Chris, it is possible this is a less computational time efficient way to structure the data. For the current capacity this will likely not be an issue, but we could always revert back to a solely numpy array structure if necessary.

Introduced some optional class init arguments to define flags for logging and image saving. This might not be implemented in the cleanest manner but works and can be optimized.

I added a executable routine and boilerplate at the bottom of the script for easier testing and debugging.