Move and organize existing analyses.
Moved and organized existing scan analyses to the new location. ScanAnalysis located within base.py file. Other more specific analyses are housed in the analyzers subdirectory.

Analyzers subdirectory partitioned by experiment to accommodate an expansion in the future.

Analysis class files intended to be written such that they can be imported or called but also with an executable for testing purposes.