Working spatial calibration for magnet deflections
Working calibration for magnetic deflections. Designed to aid e beam trajectory studies.

Mainly contained within a 'calibrations' folder as the current location is likely not the final home for this code.

Small modifications/additions made to ImageAnalysis that should merge fine but can be cleaned up in the near-term as we consolidate current code duplicates and bloat.