small change to udp_handler from geecs_python_api
Previously hardcoded to only send commands to a local instance of master control. Now, it can send commands to a MC on any computer.

note, the 'send_scan_cmd' method should probably be renamed as it can be extended to be much more general. It can be used to send any message to MC.