improved the renaming of files in scan_manager
lists of device files are sorted by shotnumber to ensure consistency.
self.scan_number_int is intialized from the scan_data argument which makes it possible to use process_and_rename() in a standalone fashion, which can be useful if an or error or something prevents this method from executing during a scan.