simple switch of the order of operations in stop_scan method of scan_manager
now the renaming of the non scalar data happens before the generation of the s files. The order switch is done for the benefit of the scan_analysis which looks for the generation of sfiles before executing analysis methods. Want to make sure the non scalar data is all fully renamed before analysis executes to avoid any race conditions.