Update scan_manager.py
parsed and reorderd the generation of the scalar data. the ScanDataScanXXX.txt file gets created first, then the process and rename happens, followed finally by creation of sfile. the process_and_rename sequence loads data from ScanDataScanXXX now instead of sfile to fix the proper workflow. Now, the creation of the sfile is a good indication of a scan ending and all other functionality should be restored