Modified how base ScanAnalyzer generates and handles the display_content file
A central yaml file is created. Each analyzer updates this file with a new entry, e.g. entry 1, entry 2, whcih can be used by the docgen module to up the experiment log. Previously, this overwiritng a single entry in the yaml, which was a bug. A.so, .yaml files are now purged (by default, but changeable) from the root analysis directory for each scan. Makes it so prior analysis can be purged.