debug how cameraImageAnalysis updates display content
the method append_display_content which is part of the base analyzer uses the file path argument to write to a file. Thus, a Path object is the wrong thing. Not sure if the append_display_content method should convert a path object or if path objects should be converted to strings first. For now, I am converting to string before using as argument