execute_scan_analysis now uploads processed data to the experiment log gdoc
an index of files is returned from each analyzer. After all analyzers have run, a flattened list of paths is sent to insert_display_content_to_doc to upload the files to a gdoc. This function uses an experiment argument that is passed to docgen.insertImageToExperimentLog which uses mapping to find the .ini file currently in use for maintaining the current experiment log. This .ini file contains the document_id needed for uploading the images to the correct file. Curently, hardcoded to insert data into rows 2-3, and cols 1-2 in a the analysis table that is part of the scan entry template.