Bugfix:  closeout actions behaving strangely
This was because I incorrectly loaded the closeout actions into the editor.  I had not updated it to load the closeout dictionary, but instead load the setup actions into the closeout action dictionary if the closeout section existed in the yaml.  This resulted in both dictionaries referencing the same `full_dictionary`, so their elements would be seemingly updated together.