moved the "try to connect to tcp" into it's own logic block
I think the logic is improved here. The checking if device is on bit, is now outside of the original try/except block for the creation of the TCPSubscriber. Now, If a TCP subscriber is created, check to see if it can connect. If not, an exception will be raised and it will return false.

note: dev_tcp.connect() does raise an error if it fails, but returns True/False in any those cases. So it can still be relied on to set Is_device_on.