small debug for handling FROG devicetype data
The FROG device type generates a an directory with the device name and two additional directories with the suffixes -Temporal and -Spatial. This is sort of like the magspectstitcher, except the primary directory for the magspecstitcher contains images, which is not the case for the FROG. So, the data for FROG is handled a little bit specially: if the device type is FROG, the primary directory gets redefined as -Temporal and the -Spatial data is dependent. This is just necessary for the fils renaming and synchornizatino of shotnumbers.