Catching error when TCP connection is refused
This occurs when trying to connect to the specified shot control device when it is not connected to GEECS.  Now it will not set Run Control but also not crash the GUI.  Still a problem if you try to start a scan using devices that are not turned on, but that's up to the user right now.

As a TODO, do a similar error handling in scan_manager to catch this and quit the scan.