Create scan_device.py
scan_device is a subclass of geecs_device. It is intended to improve native support for composite type devices/variables, by which I mean variables that are a combination of multiple devices. For example, the r56 of the HTU chicane, or a electron beam offset governed by two steering magnets.

ScanDevice is meant to handle both standard geecs devices intialized by a device name (as usual) or by a configuration defined by in a dict. There are a couple of different configurations. A set-type device is for setting variables only. It can be used in 'relative' or 'fixed' mode. A 'get' command for a set-type device does not necessairly have meaning as the dependent devices and variables are not necessarily in a state which is valid form the point of view of the defined relation.