add a functional HasoAnalysis class
This class uses the wavekit 4.3 sdk. It will check if .himg or .has files have been generated as part of the data acquisition. If .himg exists, a slopes file using the lift resolution increase method is applied. This creates a .has slopes file. From this file (or from a .has file generated during data acquisition) a phase file is generated and saved as a .tsv.