Added a "Background" scan option
Functionally equivalent to a "NoScan" (and shares much of the same code blocks).  The only difference is that if "Background" is selected then the scan's .ini file will have a `Background = "True"`, to be used for post-analysis.  We opted to make this a visible feature on the GUI rather than a more hidden "Toolbar Option" because (a) we didn't want users to forget it (b) we want to include it in multiscans (c) it doesn't increase complexity too much.