Update scan_device.py
overwritten get/set methods for GeecsDevice were returning wrong format. Now, the set return is all good. For teh get method, it's still a little funky. For non-composite, it returns the value, which will be a string or float. For composites gets, it's a little funky because of subcomponents and things like that. Could be handled better in the future probably