Update base.py
Modify auxiliary_file_path to point to sfile in analysis folder rather than Scan file in scan folder.

The scan file is designed to encapsulate the raw data copy acquired from the scan and shouldn't be touched or modified. The sfile in the analysis folder can be safely handled, appended to, etc.