Implement scan settings
Implemented functionality to define start scan and ignore scans.

I realize the GUI logic (not code logic) is not intuitive considering the use of the "Overwrite" button. Currently, the saved processed scans yaml is always overwritten so this is confusing. This works for our purposes now so will straighten this out in the future.