Finished the basic analysis routine for Rad2
Just need to gather more calibrations.  Tricky because for each VISA position we need (1) a specific cropping region (2) a specific linear fit for incoherent signal and (3) a specific calibration for the wavelength vs pixel.  This all changes at the very least when the physical setup changes, but also potentially when even the ebeam trajectory through the undulator changes.