additional logic added to data_logger to handle synchronicity
It's not super pretty yet, but the logic does appear to hold up after some testing. The _handle_TCP_message_from_device performs a lot of checks before enabling data to be logged.
After each event, it will check the standby status of an individual device, with None meaning "unknown", True meaning it is in standby and False meaning it has left standby. After each event, the overall standby status of all devices is checked until they all show true. If true, checking of indivdual statuses is skipped for subsequent calls.
Next, there is a synchronization test. Once all devices go to standby, ScanManager will fire a test/synchronization shot.  This is a hardware trigger which should cause timestamps on all devices to increment. So, again, stanby status is checked on all devices. If all have left standby, we know that the most recently acquired timestamp for each device is synchronous to the other devices and we can call that 't0' for each device and carry on taking data synchronously. However, if all devices do not successfully leave standby, the standby status for each device is set back to "None" or unkown. And the process repeats for a finite number of times. If it never gets to synchronized, the scan will exit gracefully