Implement logic for AnalysisActivator
AnalysisActivator provides logic for new dialog window, allowing for specific analyses to be toggled on and off.

ScAnalyzer logic modified to handle new button to open AnalysisActivator dialog box, return the results of user input (defining state of analysis), and update a list of available analyzers (list of AnalyzerInfo named tuples) with a status of either active or not.

The list of AnalyzerInfo named tuples is then passed to the scan_watch class upon initialization (default = None) and subsequently scan_analysis/mapping/scan_evaluator.py to be considered when evaluating the list of valid analyzers.