Added feature to import actions from save elements
The actions get stored in the list of assigned actions, and can come from either the list of setup or closeout actions.  This is additionally stored in the yaml file for assigned actions so they persist between sessions.  Slightly changed formatting in the assigned_actions.yaml and I now just access the folders using `self.main_window.app_paths`