error catching added in the message_callback of async listener
Without this error catching, any error not handled by the callback method cause this listener to close. Now, if a message isn't handled as expected, it's logged and operations continue.

For a future PR, I will consider adding a listener that checks on the status of this async_listener and restarts it if it crashes.