addition of new 'base' class for handling scan analysis of 2D array image type devices called array2D_scan_analysis
a stripped down and modified version of CameraImageAnalysis.

Goal is to remove any actual image analysis from teh ScanAnalysis class and have that all handled in image analysis. Deleted a fair bit of code from the CameraImageAnalysis and a slight refactor to make use of a new method get_image_analysis_result that uses the ImageAnalyzer class type to produce the results