ShotControl is optional, will skip synchronization
In this branch, if there is not shot control device now the scan will still start and images will be saved, but there will be no renaming of the files.  This is still useful to do some basic testing.  If we want to have a t0 for each device, will need to do more software development.