Made note about 3rd party sdk import
This was also kinda bad when it was living in image analysis.  But, I don't think it will throw any errors.  Basically, if we are running in the ScanAnalysis package then this import will know what to do.  But if we are running in GEECSScanner package then it will go to the except clause.  If the timestamp reading from filename works, then we can just axe this import statement.