first definition of the background class in image_analysis/tools/background.py
Simple object representing a background for an ImageAnalysis class. It has some basic methods to directly 'set' the background from an array or load from a file. It also has methods that can be used to construct a background from a set of images, which are particularly useful for defining background for a transient signal polluted by non homogenus background.