feat: can accommodate any number of parameters
Parameter information is now stored in a named tuple, and the base analyzer can handle a list of any number of these parameters.  It will just keep adding new axes outwards from the plot, so effectively after 3 or 4 this becomes too messy.  To plot multiple values on the same axes requires too much information for this general class, and so a custom class would need to be written that can handle it.