add new pydantic model for returns of analyzer_image.
This model will become the return type for the base image analyzer rather than a dict. It will allow for improved development/standardization without giving up any flexibility. The model accepts arbitraty entries so it can still be using a similar fashion to a dict. But, it allows for better standardization of common processing results etc.