Change return signature of analyze_image to ImageAnalyzerResult
This is a breaking change. Rather than return a dict (which is nice for flexiblity) we adopt the ImageAnalyzerResult model which allows the use of standard return type contents without sacrificing flexibility.

This also means we can demote the build_input/build_return_dictionary methods. These wre originally part of the LabView analyzer sub class, which were needed because of the highly finicky interface with labview to ensure strict type adherence. The mehtods were promoted to the base class to introduce some level of standardization across various analyzer types, but that is now superseded by the ImageAnalyzerResult model. Note, the signature of the LabviewAnalyzer.analyze_image() method is overwritten to make sure it stays compatible with the online labview analyzers.